+++ extensions/url-unshortener/src/types.tsx
@@ -2,6 +2,5 @@ export interface RedirectionStep {
   url: string;
   statusCode: number;
   statusName: string;
-  faviconUrl?: string;
   errorMessage?: string;
 }