+++ extensions/zed-recent-projects/CHANGELOG.md
@@ -1,5 +1,9 @@
 #  Zed Recent Projects Changelog
 
+## [Fixes] - 2025-07-16
+
+- Fix missing local projects
+
 ## [Fixes] - 2025-07-07
 
 - Add `Zed Dev` app build option